fix bug train list officer

+++ application/controllers/Officer/Train_list.php
@@ -17,7 +17,19 @@ class Train_list extends CI_Controller {
         }
     }
 
-    public function index(){
+    public function index($status = '')
+    {
+        if($status == 'success_delete' ){
+            $data['status']['color'] = 'success';
+            $data['status']['text'] = 'ทำการลบเรียบร้อย';
+        } else if($status == 'error_delete' ){
+            $data['status']['color'] = 'danger';
+            $data['status']['text'] = 'ผิดพลาด โปรดตรวจสอบ';
+        } 
+        else {
+            $data['status'] = '';
+        }
+
         $data['data'] = array();
         //get student has test
         foreach($this->DB_train->gets() as $row) {
@@ -30,34 +42,56 @@ class Train_list extends CI_Controller {
         $this->template->view('Officer/Train_list_view',$data);
     }
    
-    public function delete($id){
-            $this->DB_train->delete($id);
+    public function delete()
+    {
+        //check if exist
+        $this->load->library('form_validation');        
+        $this->form_validation->set_rules('id', 'id', 'trim|required|numeric');
+        
+        if ($this->form_validation->run() != FALSE) {
+            $id = $this->input->post('id');            
 
-            redirect('Officer/Train_list?delete=1');
+            if($this->DB_train->get($id)) {
+                //delete
+                $this->DB_train->delete($id);
+                return $this->index('success_delete');
+                die();
+            } else {
+                return $this->index('error_delete');
+                die();
+            }
+        } else {
+            return $this->index('error_delete');
+            die();
         }
+        
+    }
 
-        public function edit($id){
-            $data['data'] = $this->DB_train->get($id);
-            $data['train_type'] = $this->DB_train_type->gets();
-            $data['train_location'] = $this->DB_train_location->gets();
-            //list
+    public function edit($id)
+    {
+        $data['data'] = $this->DB_train->get($id);
+        $data['train_type'] = $this->DB_train_type->gets();
+        $data['train_location'] = $this->DB_train_location->gets();
+            
+        $this->template->view('Officer/Edit_Train_list_view', $data);
+    }
 
+    public function add()
+    {
+        $data['train_type'] = $this->DB_train_type->gets();
+        $data['train_location'] = $this->DB_train_location->gets();
+        $this->template->view('Officer/Add_Train_list_view', $data);
+    }
 
-            //add
-            // $data['name'] = $this->DB_train_type->update('title');
-            // $data['title'] = $this->DB_train_type->update('title');
-            // $data['lecturer'] = $this->DB_train_type->update('lecturer');
-            // $data['number_of_seat'] = $this->DB_train_type->update('number_of_seat');
-            // $data['name'] = $this->DB_train_type->update('name');
-            // $data['number_of_hour'] = $this->DB_train_type->update('number_of_hour');
+    public function post_add()
+    {
+        //insert
+    }
 
-            $this->template->view('Officer/Edit_Train_list_view', $data);
-        }
-        public function add(){
-            $this->template->view('Officer/Add_Train_list_view');
-
-    
-        }
+    public function post_edit()
+    {
+        //update
+    }
